+++ src/js/widgets.js
@@ -21,6 +21,7 @@ define( function( require, exports, module ) {
         require( '../widget/distress/distresspicker' ),
         require( '../widget/horizontal-choices/horizontalchoices' ),
         require( '../widget/analog-scale/analog-scalepicker' ),
+        require( '../widget/big-image/image-viewer' ),
     ];
 
     module.exports = widgets;