+++ CHANGELOG.md
@@ -3,10 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.5.0] - 2015-01-12
 --------------------
 ##### Added
 - Support for dynamic _required_ expressions.
+- Basic support for big-image form attributes on itext values.
 
 ##### Fixed
 - Existing XForm content of secondary external instance not properly clear if nodename is not 'root' or if multiple root-level nodes are present.